Sidekiq fix for all

+++ db/migrate/20260916091123_remove_details_from_client.rb
@@ -1,7 +1,7 @@
 class RemoveDetailsFromClient < ActiveRecord::Migration[8.1]
   def change
-    remove_column :clients, :website, :string
-    remove_column :clients, :settlement_bank, :string
-    remove_column :clients, :tax_pin, :string
+    remove_column :clients, :website, :string if exists_column?(:clients, :website)
+    remove_column :clients, :settlement_bank, :string if exists_column?(:clients, :settlement_bank)
+    remove_column :clients, :tax_pin, :string if exists_column?(:clients, :tax_pin)
   end
 end
